tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.quadraticCurveTo.scaled.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.quadraticCurveTo.scaled</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.quadraticCurveTo.scaled</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.scale(1000, 1000);
  ctx.strokeStyle = '#0f0';
  ctx.lineWidth = 0.055;
  ctx.beginPath();
  ctx.moveTo(-1, 1.05);
  ctx.quadraticCurveTo(0, -1, 1.2, 1.05);
  ctx.stroke();
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 98,1, 0,255,0,255);
  _assertPixel(canvas, 1,48, 0,255,0,255);
  _assertPixel(canvas, 98,48, 0,255,0,255);
  t.done();

});
</script>
